add license explain
