--- fql/src/main/java/cn/fusiondb/core/Ast.java
@@ -1,5 +1,0 @@
-package cn.fusiondb.core;
-
-public class Ast {
-
-}
