--- fql/src/main/scala/cn/fusiondb/core/execution/command/commands.scala
@@ -1,7 +1,0 @@
-package cn.fusiondb.core.execution.command
-
-trait commands {
-
-  def output: Seq[AnyRef] = Seq.empty
-
-}
