--- fql/src/main/scala/cn/fusiondb/core/parser/FqlParse.scala
@@ -1,10 +1,0 @@
-package cn.fusiondb.core.parser
-
-import org.apache.spark.sql.internal.SQLConf
-
-class FqlParse(conf: SQLConf)  extends AbstractSqlParser {
-
-  val astBuilder = new AstBuilder(conf)
-
-  // sql parser
-}
