+++ fql/src/main/scala/cn/fusiondb/fql/execution/command/RunnableCommand.scala
@@ -1,0 +1,40 @@
+/*
+ * Copyright 2019 Fusionlab, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package cn.fusiondb.fql.execution.command
+
+import org.apache.spark.sql.{Row, SparkSession}
+
+/**
+  *
+  * Create by xujiang on 2018/12/29
+  *
+  */
+trait RunnableCommand extends commands {
+
+  override def output = super.output
+
+  def run(sparkSession: SparkSession): Seq[Row]
+}
+
+/**
+  *
+  * @param cmd
+  */
+case class ExecutedCommandExec(cmd: RunnableCommand) {
+
+  def output: Seq[AnyRef] = cmd.output
+
+}
